TST/REF: misplaced tests.internals tests
